Migration step 2: backfill data and read from new column
In this second step, we backfill the data from the old column to the new
column, and we update the code to read from the new column.

We stop all writing to the old column and mark the field as `virtual`:

- We need to stop all writing to the old column so that we can safely
  remove the column in the next deploy. If anything tries to write to
  the column when it's removed, it'll break.

- We need to mark the field as `virtual` so that it's no longer expected
  to read from it. If records try to load a `User` with the "old" code
  after the next deployment, then it would fail because we've removed
  the column. So we mark it as `virtual` so it remains in the struct for
  now but not tied to a colum.

Note that we do not remove the old column/field in this step because
when we deploy this work, there will be instances running code that use
the old field/column.